engine_setup: Add pki link for AnsibleRunner
