+++ packaging/setup/plugins/ovirt-engine-setup/ovirt-engine/pki/runner.py
@@ -1,0 +1,54 @@
+#
+# ovirt-engine-setup -- ovirt engine setup
+#
+# Copyright oVirt Authors
+# SPDX-License-Identifier: Apache-2.0
+#
+#
+
+
+"""Apache ansible-runner plugin."""
+
+
+import gettext
+import textwrap
+import os
+
+from otopi import constants as otopicons
+from otopi import filetransaction
+from otopi import plugin
+from otopi import util
+
+from ovirt_engine_setup import constants as osetupcons
+from ovirt_engine_setup.engine import constants as oenginecons
+from ovirt_engine_setup.engine_common import constants as oengcommcons
+
+
+def _(m):
+    return gettext.dgettext(message=m, domain='ovirt-engine-setup')
+
+
+@util.export
+class Plugin(plugin.PluginBase):
+    """ansible-runner plugin"""
+
+    def __init__(self, context):
+        super(Plugin, self).__init__(context=context)
+
+    @plugin.event(
+        stage=plugin.Stages.STAGE_MISC,
+        condition=lambda self: self.environment[
+            oenginecons.CoreEnv.ENABLE
+        ],
+    )
+    def _misc(self):
+        if not os.path.exists(
+            oengcommcons.FileLocations.OVIRT_ENGINE_PKI_ANSIBLE_CA_CERT
+        ):
+            os.symlink(
+                oenginecons.FileLocations.OVIRT_ENGINE_PKI_ENGINE_CA_CERT,
+                oengcommcons.FileLocations.OVIRT_ENGINE_PKI_ANSIBLE_CA_CERT
+            )
+
+
+# vim: expandtab tabstop=4 shiftwidth=4
